+++ source/style/twitterFirstMinute.js
@@ -1,6 +1,7 @@
 import { style } from './twitter'
 
-const steps = style.steps.slice(1)
+// Skip "seconds", "now", "seconds" (from `0.5`) steps.
+const steps = style.steps.slice(3)
 steps[0] = {
 	...steps[0],
 	// Starts showing `1m` after 45.5 seconds.